LayoutTests/imported/w3c/web-platform-tests/svg/animations/non-additive-type-from-by-animation-expected.txt


PASS This is a from by animation for all non-additive property types - should have no effect.
